backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Database and ORM - Python 3.13 compatible
asyncpg>=0.30.0  # Updated for Python 3.13 compatibility
psycopg2-binary>=2.9.0  # PostgreSQL adapter for SQLAlchemy
databases[postgresql]>=0.9.0  # Updated for SQLAlchemy 2.x compatibility
sqlalchemy>=2.0.23
alembic>=1.12.1

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Client - Compatible with supabase
httpx>=0.24.1,<0.25.0  # Compatible with supabase requirement
aiohttp>=3.9.1

# Data Processing - Python 3.13 compatible
pandas>=2.2.0  # Updated for Python 3.13 compatibility
numpy>=2.0.0  # Updated for Python 3.13 compatibility
pydantic>=2.10.0  # Updated for Python 3.13 compatibility
pydantic-settings>=2.6.0  # Updated for Python 3.13 compatibility

# External Service Integrations
supabase==2.0.2  # Fixed version to avoid compatibility issues
pytrends>=4.9.2
beautifulsoup4>=4.12.2
openai>=1.3.7
shopifyapi>=12.0.0  # Shopify API client

# Azure Services
azure-keyvault-secrets>=4.7.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0

# Background Jobs and Caching
celery>=5.3.4
redis>=5.0.1
python-redis-lock>=4.0.0

# Monitoring and Logging
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0
prometheus-client>=0.19.0

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1
pre-commit>=3.6.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.8

# Debugging
ipdb>=0.13.13

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
email-validator